refactoring package for detail activity view
